Pod should match the inferGroup revision instead of modelInfer revision

+++ pkg/infer-controller/controller/modelinfer_controller.go
@@ -213,15 +213,20 @@ func (c *ModelInferController) updatePod(oldObj, newObj interface{}) {
 		}
 		return
 	}
-	revision := utils.HashModelInferRevision(mi.Spec.Template.Roles)
+
+	if c.shouldSkipPodHandling(mi, inferGroupName, newPod) {
+		// Pod revision mismatch inferGroup, this can rarely happen
+		return
+	}
+
 	switch {
-	case utils.IsPodRunningAndReady(newPod) && utils.CheckPodRevision(newPod, revision):
+	case utils.IsPodRunningAndReady(newPod):
 		// The pod is available, that is, the state is running, and the container is ready
-		err = c.handleReadyPod(mi, inferGroupName, newPod, revision)
+		err = c.handleReadyPod(mi, inferGroupName, newPod)
 		if err != nil {
 			klog.Errorf("handle running pod failed: %v", err)
 		}
-	case utils.IsPodFailed(newPod) || utils.ContainerRestarted(newPod) || !utils.CheckPodRevision(newPod, revision):
+	case utils.IsPodFailed(newPod) || utils.ContainerRestarted(newPod):
 		// handleErrorPod is not called until modelInfer has been called.
 		if !c.initialSync {
 			return
@@ -252,6 +257,9 @@ func (c *ModelInferController) deletePod(obj interface{}) {
 
 	mi, inferGroupName, err := c.getModelInfer(pod)
 	if err == nil {
+		if c.shouldSkipPodHandling(mi, inferGroupName, pod) {
+			return
+		}
 		c.DeleteInferGroup(mi, inferGroupName)
 		return
 	}
@@ -314,7 +322,7 @@ func (c *ModelInferController) syncModelInfer(ctx context.Context, key string) e
 		return err
 	}
 	// only fields in roles can be modified in rolling updates
-	revision := utils.HashModelInferRevision(mi.Spec.Template.Roles)
+	revision := utils.Revision(mi.Spec.Template.Roles)
 
 	err = c.manageReplicas(ctx, mi, revision)
 	if err != nil {
@@ -608,19 +616,19 @@ func (c *ModelInferController) manageRollingUpdate(mi *workloadv1alpha1.ModelInf
 		return fmt.Errorf("cannot get inferGroupList from store, err:%v", err)
 	}
 	// we terminate the inferGroup with the largest ordinal that does not match the update revision.
-	for target := len(inferGroupList) - 1; target >= updateMin; target-- {
-		if c.isInferGroupOutdated(inferGroupList[target], mi.Namespace, revision) {
+	for i := len(inferGroupList) - 1; i >= updateMin; i-- {
+		if c.isInferGroupOutdated(inferGroupList[i], mi.Namespace, revision) {
 			// target inferGroup is not the latest version, needs to be updated
-			klog.V(2).Infof("inferGroup %s will be terminating for update", inferGroupList[target].Name)
-			c.DeleteInferGroup(mi, inferGroupList[target].Name)
+			klog.V(2).Infof("inferGroup %s will be terminating for update", inferGroupList[i].Name)
+			c.DeleteInferGroup(mi, inferGroupList[i].Name)
 			return nil
 		}
-		if inferGroupList[target].Status != datastore.InferGroupRunning {
+		if inferGroupList[i].Status != datastore.InferGroupRunning {
 			// target inferGroup is the latest version, but not running. We need to wait for the status to change to running.
 			// If the group fails after rolling, it will automatically be deleted and rebuilt when detecting the pod failure.
 			// If the group still pending due to reasons such as being unable to be scheduled, rolling update process will stop
 			// to avoid affecting other groups that are running normally.
-			klog.V(4).Infof("waiting for the infergroup %s status become running", inferGroupList[target].Name)
+			klog.V(4).Infof("waiting for the infergroup %s status become running", inferGroupList[i].Name)
 			return nil
 		}
 		// target inferGroup is already the latest version and running, processing the rolling update of the next group.
@@ -629,12 +637,12 @@ func (c *ModelInferController) manageRollingUpdate(mi *workloadv1alpha1.ModelInf
 	return nil
 }
 
-func (c *ModelInferController) handleReadyPod(mi *workloadv1alpha1.ModelInfer, inferGroupName string, newPod *corev1.Pod, revision string) error {
+func (c *ModelInferController) handleReadyPod(mi *workloadv1alpha1.ModelInfer, inferGroupName string, newPod *corev1.Pod) error {
 	// Add the running pod to the global storage and try to update the inferGroup status
 	c.store.AddRunningPodToInferGroup(types.NamespacedName{
 		Namespace: mi.Namespace,
 		Name:      mi.Name,
-	}, inferGroupName, newPod.Name, revision)
+	}, inferGroupName, newPod.Name, utils.PodRevision(newPod))
 	ready, err := c.checkInferGroupReady(mi, inferGroupName)
 	if err != nil {
 		return fmt.Errorf("failed to check inferGroup status, err: %v", err)
@@ -735,7 +743,7 @@ func (c *ModelInferController) checkInferGroupReady(mi *workloadv1alpha1.ModelIn
 	return true, nil
 }
 
-func (c *ModelInferController) isInferGroupOutdated(group datastore.InferGroup, namespace string, newHash string) bool {
+func (c *ModelInferController) isInferGroupOutdated(group datastore.InferGroup, namespace, newRevision string) bool {
 	// Find the pods corresponding to infergroup
 	req, err := labels.NewRequirement(workloadv1alpha1.GroupNameLabelKey, selection.Equals, []string{group.Name})
 	if err != nil {
@@ -754,12 +762,7 @@ func (c *ModelInferController) isInferGroupOutdated(group datastore.InferGroup, 
 	}
 	// Check all pods match the newHash
 	for _, pod := range pods {
-		revision, ok := pod.Labels[workloadv1alpha1.RevisionLabelKey]
-		if !ok {
-			klog.V(2).Infof("pod of inferGroup %s cannot find revision label", group.Name)
-			return true
-		}
-		if revision != newHash {
+		if utils.PodRevision(pod) != newRevision {
 			return true
 		}
 	}
@@ -776,4 +779,20 @@ func (c *ModelInferController) getModelInfer(pod *corev1.Pod) (*workloadv1alpha1
 		return nil, "", err
 	}
 	return mi, inferGroupName, nil
+}
+
+// shouldSkipPodHandling checks if a pod should be skipped based on revision mismatch
+func (c *ModelInferController) shouldSkipPodHandling(mi *workloadv1alpha1.ModelInfer, inferGroupName string, pod *corev1.Pod) bool {
+	podRevision := utils.PodRevision(pod)
+	inferGroup := c.store.GetInferGroup(types.NamespacedName{
+		Namespace: mi.Namespace,
+		Name:      mi.Name,
+	}, inferGroupName)
+	if inferGroup != nil && inferGroup.Revision != podRevision {
+		// If the pod revision is not equal to the inferGroup revision, we do not need to handle it.
+		klog.V(4).Infof("pod %s/%s revision %s is not equal to inferGroup %s revision %s, skip handling",
+			pod.Namespace, pod.Name, podRevision, inferGroupName, inferGroup.Revision)
+		return true
+	}
+	return false
 }
